third_party/blink/web_tests/accessibility/accessibility-node-memory-management-expected.txt

This test makes sure that AccessibilityNodeObjects are properly detached when the node they point to is deleted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS expectedButtonRole != expectedDetachedRole is true
PASS canvasButtonRole is expectedButtonRole
PASS detachedCanvasButtonRole is expectedDetachedRole
PASS successfullyParsed is true

TEST COMPLETE
